Create basic point cloud scanning setup
This change introduces a basic point cloud scanner, which can export to a single file. It is important to delete that file from a device before starting a new scan, since points will be appended to it. The performance is currently quite bad with low FPS when points are extracted from a frame.

The point cloud file is provided in ASCII format, as a .xyz file. It is a CSV with one point per line and each point with the values x,y,z,r,g,b,confidence.

The coordinates are only the ARCore World Coordinates (not global coordinates), so georeferencing isn't included yet. An initial test has been done to use Lat/Long/Altitude instead of x/y/z, but this was not easily visualized in CloudCompare, which is why the x/y/z coordinates are used for now.

To generate the point cloud the depth image is processed pixel by pixel. If the normalized confidence value for a given depth pixel is below a threshold (0.1 for now, but this will be increased later to exclude bad pixels early and save processing time) the pixel will be ignored. If a pixel has a high enough confidence value, its world coordinates are calculates with the camera intrinsics of the depth image. In addition, the regular camera image is converted to RGB using BoofCV and the color for a given depth pixel is retrieved from that BoofCV RGB image.

The SceneView ARCore version is overridden with a newer one to benefit from better performance and helper methods related to the Geospatial API.

Two extension functions are created to allow easy configuration of fullscreen mode.

A status text is added to the AR view to show the current status of the Geospatial API.