Make Scans configurable
Creates a new structure of the app flow. On start, a list of existing scans is displayed. A new one can be created with a FloatingActionButton. Each scan holds the scanning configuration in a meta.json file, and the same folder will contain the point cloud data in the file raw.xyz.

The goal of this new structure is to allow multiple stages of post-processing of scans while retaining the original data to test different processing approaches without having to generate new data every time. Also, since Scans can now be named, the resulting data is easier to navigate.

This change also introduces some better UI elements to make the current scanning status more visible.

Finally, a crude way to automatically assign UTM coordinates to points is added. Currently, this produces striping artifacts in the resulting point cloud probably due to rounding or calculation inaccuracies. This should be addressed in the future. This coordinate conversion is currently done during the scan and replaces the normal coordinates, but later on this should be converted to a post-processing step.