Refactor to cleaner architecture and accelerate processing
This commit mainly focuses on cleaning up the code and refactoring to a cleaner architecture.

The processing of frames is now done in a dedicated class as opposed to within `FramePointCloud` objects.

The new class `FrameImageData` manages access to the three images that have to be handled each frame (Depth, Confidence and RGB) and does some processing to allow easy access to each pixel's values.