Introduce postprocessing architecture and UTM postprocessor
This commit introduces an architecture for postprocessing of the raw scans. The architecture allows for easy creation of configurable postprocessing steps.

Scans now track their total point count.

Framerates during a scan are improved by removing calls to `String.format` in the `TimingHelper`.

A `BaseFragment` is introduced to centralize handling of viewBinding.