Properly store lat/long data
Previously, the data for Latitude and Longitude was stored in Floats, which resulted in big precision losses once conversion to UTM was done. The data is now stored as doubles. Also, a class for converting from Camerapose to Pointpose is introduced to allow this calculation to take place as a postprocessing step (at a later date).