Extract georeferencing from runtime-processing
Previously, the geocoordinates of points were calculated during the scan itself. This however costs valuable framerate and reduces point count and tracking quality.

With this change only the camera pose and camera geocoordinates are stored now once per frame. The UTM postprocessor then uses that data to calculate the global position of each point after the scan.

Postprocessing now uses Strings instead of PixelData objects to allow for more flexibility down the line.

Also, local positions of the camera are now in reference to the Geoanchor, which was previously ignored. The Geoanchor is oriented without and rotation, so any position within its coordinate system is automatically in the East-Up-South coordinate system.

The input to the CTS library for converting between lat/long and UTM had to be swapped for valid coordinate results.